fix some stan errors
