+++ packages/address/config/address.php
@@ -4,7 +4,6 @@ use Moox\Address\Models\Address;
 use Moox\Address\Models\Addressable;
 use Moox\Category\Resources\CategoryResource;
 use Moox\Media\Resources\MediaResource;
-use Moox\News\Moox\Entities\News\News\NewsResource;
 use Moox\Tag\Resources\TagResource;
 use Moox\User\Models\User;
 use Moox\User\Resources\UserResource;
@@ -78,9 +77,6 @@ return [
 
             'scopes' => [
                 'allowed' => [
-                    'news' => [
-                        'resource' => NewsResource::class,
-                    ],
                     'media' => [
                         'resource' => MediaResource::class,
                     ],
