+++ packages/address/database/Factories/AddressFactory.php
@@ -28,10 +28,10 @@ class AddressFactory extends Factory
             ]),
             'name' => fake()->company(),
             'street' => fake()->streetName().' '.fake()->buildingNumber(),
-            'street2' => fake()->optional(0.15)->secondaryAddress(),
+            'street2' => fake()->optional(0.15)->streetAddress(),
             'postal_code' => fake()->postcode(),
             'city' => fake()->city(),
-            'state' => fake()->optional(0.4)->state(),
+            'state' => fake()->optional(0.4)->randomElement(['BE', 'BY', 'HH', 'NW', 'HE']),
             'country_code' => fake()->countryCode(),
             'is_primary' => false,
             'data' => null,
