+++ packages/address/src/Resources/Address/Pages/EditAddress.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Moox\Address\Resources\Address\Pages;
 
+use Illuminate\Database\Eloquent\Model;
+use Moox\Address\Exceptions\DuplicateAddressException;
 use Moox\Address\Resources\Address\Pages\Concerns\HandlesDuplicateAddressValidation;
 use Moox\Address\Resources\Address\Pages\Concerns\InitializesValidationBag;
 use Moox\Address\Resources\AddressResource;
@@ -15,4 +17,13 @@ class EditAddress extends BaseEditRecord
     use InitializesValidationBag;
 
     protected static string $resource = AddressResource::class;
+
+    protected function handleRecordUpdate(Model $record, array $data): Model
+    {
+        try {
+            return parent::handleRecordUpdate($record, $data);
+        } catch (DuplicateAddressException $exception) {
+            throw $this->duplicateAddressExceptionForForm($exception);
+        }
+    }
 }
